LayoutTests/http/tests/local/fileapi/file-last-modified-after-delete-expected.txt

Test File.lastModifiedDate returns null if the information is not available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.dataTransfer contains a File object on drop.
PASS lastModifiedDate is null
PASS successfullyParsed is true

TEST COMPLETE
